refactor: move tests to correct class
